Move Market.tsx out of MarketDetailsBody/ folder
